+++ Assets/Plugins/Dash/Runtime/Scripts/RuntimeCore/Node/Nodes/Event/SendCustomEventNode.cs
@@ -34,7 +34,7 @@ namespace Dash
                     DashRuntimeCore.Instance.SendEvent(eventName, sendData ? p_flowData : NodeFlowDataFactory.Create());
                 }
                 #else
-                DashCore.Instance.SendEvent(eventName, sendData ? p_flowData : NodeFlowDataFactory.Create());
+                DashRuntimeCore.Instance.SendEvent(eventName, sendData ? p_flowData : NodeFlowDataFactory.Create());
                 #endif
             }
             else